更改用户画面的编辑和删除后的返回值类型 直接跳转到列表页

+++ src/main/java/org/yidu/novel/action/user/MessageAction.java
@@ -2,7 +2,9 @@ package org.yidu.novel.action.user;
 
 import java.util.List;
 
+import org.apache.struts2.convention.annotation.Result;
 import org.springframework.transaction.annotation.Transactional;
+import org.yidu.novel.action.base.AbstractBaseAction;
 import org.yidu.novel.action.base.AbstractUserBaseAction;
 import org.yidu.novel.bean.MessageSearchBean;
 import org.yidu.novel.constant.YiDuConstants;
@@ -24,6 +26,16 @@ public class MessageAction extends AbstractUserBaseAction {
      * 串行化版本统一标识符
      */
     private static final long serialVersionUID = 6707140588808286899L;
+
+    /**
+     * 功能名称。
+     */
+    public static final String NAME = "message";
+
+    /**
+     * URL。
+     */
+    public static final String URL = NAMESPACE + "/" + NAME;
 
     private int messageno;
 
@@ -56,6 +68,11 @@ public class MessageAction extends AbstractUserBaseAction {
     }
 
     @Override
+    public String getBackUrl() {
+        return MessageAction.URL;
+    }
+
+    @Override
     protected void loadData() {
         MessageSearchBean searchBean = new MessageSearchBean();
         int userno = LoginManager.getLoginUser().getUserno();
@@ -68,15 +85,14 @@ public class MessageAction extends AbstractUserBaseAction {
     public String delete() {
         if (messageno != 0) {
             TMessage message = this.messageService.getByNo(messageno);
-            if (message.getUserno() == LoginManager.getLoginUser().getUserno()) {
+            if (message != null && message.getUserno() == LoginManager.getLoginUser().getUserno()) {
                 this.messageService.deleteByNo(messageno);
             } else {
                 addActionError(getText("errors.unauthority.message"));
                 return FREEMARKER_ERROR;
             }
         }
-        this.loadData();
-        return FREEMARKER;
+        return REDIRECT;
     }
 
 }
